NPCs are loaded in, display, and collide.
NPC file was deleted and npc info was moved to the
tile file as npcs are now tiles. The room.py
now loads in npc json objects from dictionaries
and creates them as tile objects. The controller
checks for npcs in addition to collidable objects.
The view file draws npcs in their own tile group.